requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
langchain
langchain-community
